Shuffled functions into common folder
