+++ clients/rocblascommon/utility.cpp
@@ -43,37 +43,37 @@ std::string rocblas_exepath()
     return pathstr;
 }
 
-/* ============================================================================================
- */
-/*  timing:*/
+// /* ============================================================================================
+//  */
+// /*  timing:*/
 
-/*! \brief  CPU Timer(in microsecond): synchronize with the default device and
- * return wall time */
-double get_time_us()
-{
-    hipDeviceSynchronize();
-    struct timespec tv;
-    clock_gettime(CLOCK_MONOTONIC, &tv);
-    return tv.tv_sec * 1'000'000llu + (tv.tv_nsec + 500llu) / 1000;
-}
+// /*! \brief  CPU Timer(in microsecond): synchronize with the default device and
+//  * return wall time */
+// double get_time_us()
+// {
+//     hipDeviceSynchronize();
+//     struct timespec tv;
+//     clock_gettime(CLOCK_MONOTONIC, &tv);
+//     return tv.tv_sec * 1'000'000llu + (tv.tv_nsec + 500llu) / 1000;
+// }
 
-/*! \brief  CPU Timer(in microsecond): synchronize with given queue/stream and
- * return wall time */
-double get_time_us_sync(hipStream_t stream)
-{
-    hipStreamSynchronize(stream);
-    struct timespec tv;
-    clock_gettime(CLOCK_MONOTONIC, &tv);
-    return tv.tv_sec * 1'000'000llu + (tv.tv_nsec + 500llu) / 1000;
-}
+// /*! \brief  CPU Timer(in microsecond): synchronize with given queue/stream and
+//  * return wall time */
+// double get_time_us_sync(hipStream_t stream)
+// {
+//     hipStreamSynchronize(stream);
+//     struct timespec tv;
+//     clock_gettime(CLOCK_MONOTONIC, &tv);
+//     return tv.tv_sec * 1'000'000llu + (tv.tv_nsec + 500llu) / 1000;
+// }
 
-/*! \brief  CPU Timer(in microsecond): no GPU synchronization */
-double get_time_us_no_sync()
-{
-    struct timespec tv;
-    clock_gettime(CLOCK_MONOTONIC, &tv);
-    return tv.tv_sec * 1'000'000llu + (tv.tv_nsec + 500llu) / 1000;
-}
+// /*! \brief  CPU Timer(in microsecond): no GPU synchronization */
+// double get_time_us_no_sync()
+// {
+//     struct timespec tv;
+//     clock_gettime(CLOCK_MONOTONIC, &tv);
+//     return tv.tv_sec * 1'000'000llu + (tv.tv_nsec + 500llu) / 1000;
+// }
 
 /* ============================================================================================
  */
