+++ library/src/rocsolver_logger.cpp
@@ -320,50 +320,3 @@ catch(...)
     return exception_to_rocblas_status();
 }
 }
-
-/***************************************************************************
- * rocsolver_logger member functions
- ***************************************************************************/
-template <>
-char rocsolver_logger::get_precision<float>()
-{
-    return 's';
-}
-template <>
-char rocsolver_logger::get_precision<double>()
-{
-    return 'd';
-}
-template <>
-char rocsolver_logger::get_precision<rocblas_float_complex>()
-{
-    return 'c';
-}
-template <>
-char rocsolver_logger::get_precision<rocblas_double_complex>()
-{
-    return 'z';
-}
-
-double rocsolver_logger::get_time_us()
-{
-    hipDeviceSynchronize();
-    struct timespec tv;
-    clock_gettime(CLOCK_MONOTONIC, &tv);
-    return tv.tv_sec * 1'000'000llu + (tv.tv_nsec + 500llu) / 1000;
-}
-
-double rocsolver_logger::get_time_us_sync(hipStream_t stream)
-{
-    hipStreamSynchronize(stream);
-    struct timespec tv;
-    clock_gettime(CLOCK_MONOTONIC, &tv);
-    return tv.tv_sec * 1'000'000llu + (tv.tv_nsec + 500llu) / 1000;
-}
-
-double rocsolver_logger::get_time_us_no_sync()
-{
-    struct timespec tv;
-    clock_gettime(CLOCK_MONOTONIC, &tv);
-    return tv.tv_sec * 1'000'000llu + (tv.tv_nsec + 500llu) / 1000;
-}
